requirements.txt
google-api-python-client
google-auth
kraken
numpy
opencv-python
pandas
pillow
pytesseract
tqdm
watchdog
